Backend/requirements.txt
Flask>=2.0
Flask-RESTful>=0.3.9
Flask-JWT-Extended>=4.0
Flask-Security-Too>=4.0
Flask-Mail>=0.9
Flask-Migrate>=3.1
Flask-SQLAlchemy>=3.0
Flask-Bcrypt>=1.0
Flask-Caching>=2.0
Flask-Cors>=3.0
celery>=5.2
redis>=4.0
SQLAlchemy>=1.4
alembic>=1.10
reportlab>=4.0
python-dotenv>=1.0
gunicorn>=20.1
psycopg2-binary>=2.9
requests>=2.28
flask-socketio>=5.3
eventlet>=0.33
python-socketio>=5.8
python-engineio>=4.5
qrcode>=7.3
Pillow>=9.0